internal_russh_forked_ssh_key

Enum Error

#[non_exhaustive]
pub enum Error {

    AlgorithmUnknown,
    AlgorithmUnsupported {
        algorithm: Algorithm,
    },
    CertificateFieldInvalid(Field),
    CertificateValidation,
    Crypto,
    Decrypted,
    Ecdsa(Error),
    Encoding(Error),
    Encrypted,
    FormatEncoding,
    Io(ErrorKind),
    Namespace,
    PublicKey,
    Ppk(PpkParseError),
    Time,
    TrailingData {
        remaining: usize,
    },
    Version {
        number: u32,
    },
    MpintEncoding,

}

Error type.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AlgorithmUnknown

Unknown algorithm.

This is returned when an algorithm is completely unknown to this crate.

AlgorithmUnsupported

Unsupported algorithm.

This is typically returned when an algorithm is recognized, but the relevant crate features to support it haven’t been enabled.

It may also be returned in the event an algorithm is inappropriate for a given usage pattern or context.

Fields

algorithm: Algorithm

Algorithm identifier.

CertificateFieldInvalid(Field)

Certificate field is invalid or already set.

CertificateValidation

Certificate validation failed.

Crypto

Cryptographic errors.

Decrypted

Cannot perform operation on decrypted private key.

Ecdsa(Error)

ECDSA key encoding errors.

Encoding(Error)

Encoding errors.

Encrypted

Cannot perform operation on encrypted private key.

FormatEncoding

Other format encoding errors.

Io(ErrorKind)

Input/output errors.

Namespace

Namespace invalid.

PublicKey

Public key is incorrect.

Ppk(PpkParseError)

PuTTY format parsing errors.

Time

Invalid timestamp (e.g. in a certificate)

TrailingData

Unexpected trailing data at end of message.

Fields

remaining: usize

Number of bytes of remaining data at end of message.

Version

Unsupported version.

Fields

number: u32

Version number.

MpintEncoding